+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drivers.java
@@ -1,20 +1,32 @@
 package org.firstinspires.ftc.teamcode;
+
 import static org.firstinspires.ftc.teamcode.UtilClass.ServoUtil.closeClaw;
 import static org.firstinspires.ftc.teamcode.UtilClass.ServoUtil.openClaw;
-import static org.firstinspires.ftc.teamcode.UtilClass.ServoUtil.retractAirplane;
-import static org.firstinspires.ftc.teamcode.UtilClass.ServoUtil.shootAirplane;
-import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.*;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.airplaneArmed;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.cRE;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.claw1;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.claw2;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.flipperServo;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.intakeMax;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.intakeMin;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.intakePower;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.liftMax;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.liftMin;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.liftPower;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.slideMax;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.slideMin;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.slidePower;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.slowModeIsOn;
+import static org.firstinspires.ftc.teamcode.opModes.HardwareConfig.timer;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.teamcode.UtilClass.Blink;
 import org.firstinspires.ftc.teamcode.UtilClass.ServoUtil;
 import org.firstinspires.ftc.teamcode.ggutil.pure.fromTutorial.com.company.Range;
-import org.firstinspires.ftc.teamcode.opModes.HardwareConfig;
 
 public class Drivers {
-    public static final String[] driverControls = {"Chase", "Camden", "Kian", "Grady", "Michael"}, otherControls = driverControls;
+    public static final String[] driverControls = {"Chase", "Camden", "Kian", "Grady", "Michael","Graden"}, otherControls = driverControls;
     public static final int baseDriver = 0, baseOther = 1;//list integer of base driver and other controls
     public static int dIndex = baseDriver, oIndex = baseOther;//list integer of driver and other controls
     public static String currDriver = driverControls[dIndex], currOther = otherControls[oIndex];//list string of driver and other controls
@@ -24,13 +36,13 @@ public class Drivers {
     public static boolean dDownHigh = false;
 
     public static void bindDriverButtons(OpMode myOpMode) {
-        //"Chase", "Camden", "Kian", "Grady", "Michael"
+        //"Chase", "Camden", "Kian", "Grady", "Michael","Graden"
         if (currDriver == driverControls[0]) {//Chase
             fieldCentric = false;
             //slowmode
             if (myOpMode.gamepad1.dpad_down && !dDownHigh && !slowModeIsOn) {
                 slowModeIsOn = true;
-            }else if (myOpMode.gamepad1.dpad_down && !dDownHigh && slowModeIsOn){
+            } else if (myOpMode.gamepad1.dpad_down && !dDownHigh && slowModeIsOn) {
                 slowModeIsOn = false;
             }
             dDownHigh = myOpMode.gamepad1.dpad_down;
@@ -44,7 +56,7 @@ public class Drivers {
             //slowmode
             if (myOpMode.gamepad1.dpad_down && !dDownHigh && !slowModeIsOn) {
                 slowModeIsOn = true;
-            }else if (myOpMode.gamepad1.dpad_down && !dDownHigh && slowModeIsOn){
+            } else if (myOpMode.gamepad1.dpad_down && !dDownHigh && slowModeIsOn) {
                 slowModeIsOn = false;
             }
             dDownHigh = myOpMode.gamepad1.dpad_down;
@@ -55,142 +67,10 @@ public class Drivers {
         if (currDriver == driverControls[4]) {//Michael
             fieldCentric = false;
         }
-    }
-
-    public static void bindOtherButtons(OpMode myOpMode){
-        //"Chase", "Camden", "Kian", "Grady", "Michael"
-        if (!airplaneArmed && timer.seconds() > 80) {
-            airplaneArmed = true;
-            myOpMode.gamepad2.runRumbleEffect(cRE);
-        }
-        Blink.selectLights(myOpMode);
-        if (currOther == otherControls[0]) {//Chase
-//            if (myOpMode.gamepad2.circle && airplaneArmed){
-//                airplanePower = airplaneMax;
-//            }else{
-//                airplanePower = 0;
-//            }
-//            if (myOpMode.gamepad2.dpad_up){
-//                shootAirplane(airplaneServo);
-//            }else if (myOpMode.gamepad2.dpad_down){
-//                retractAirplane(airplaneServo);
-//            }
-            if (myOpMode.gamepad2.dpad_up){
-                openClaw(claw1);
-            }
-            if (myOpMode.gamepad2.dpad_down){
-                closeClaw(claw1);
-            }
-            if (myOpMode.gamepad2.triangle){
-                openClaw(claw2);
-            }
-            if (myOpMode.gamepad2.cross){
-                closeClaw(claw2);
-            }
-            liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, liftMin, liftMax);
-        }
-        if (currOther == otherControls[1]) {//Camden
-//            if (myOpMode.gamepad2.cross && airplaneArmed){
-//                airplanePower = airplaneMax;
-//            }else{
-//                airplanePower = 0;
-//            }
-//            if (myOpMode.gamepad2.dpad_up){
-//                shootAirplane(airplaneServo);
-//            }else if (myOpMode.gamepad2.dpad_down){
-//                retractAirplane(airplaneServo);
-//            }
-            if (myOpMode.gamepad2.right_bumper){
-                openClaw(claw1);
-            }
-            if (myOpMode.gamepad2.left_bumper){
-                closeClaw(claw1);
-            }
-            if (myOpMode.gamepad2.dpad_down){
-                openClaw(claw2);
-            }
-            if (myOpMode.gamepad2.dpad_up){
-                closeClaw(claw2);
-            }
-            liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, liftMin, liftMax);
-        }
-        if (currOther == otherControls[2]) {//Kian
-//            if (myOpMode.gamepad2.circle && airplaneArmed){
-//                airplanePower = airplaneMax;
-//            }else{
-//                airplanePower = 0;
-//            }
-//            if (myOpMode.gamepad2.dpad_up){
-//                shootAirplane(airplaneServo);
-//            }else if (myOpMode.gamepad2.dpad_down){
-//                retractAirplane(airplaneServo);
-//            }
-            if (myOpMode.gamepad2.right_bumper){
-                openClaw(claw1);
-            }
-            if (myOpMode.gamepad2.left_bumper){
-                closeClaw(claw1);
-            }
-            if (myOpMode.gamepad2.dpad_right){
-                openClaw(claw2);
-            }
-            if (myOpMode.gamepad2.dpad_left){
-                closeClaw(claw2);
-            }
-            liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, liftMin, liftMax);
-        }
-        if (currOther == otherControls[3]) {//Grady
-//            if (myOpMode.gamepad2.circle && airplaneArmed){
-//                airplanePower = airplaneMax;
-//            }else{
-//                airplanePower = 0;
-//            }
-//            if (myOpMode.gamepad2.dpad_up){
-//                shootAirplane(airplaneServo);
-//            }else if (myOpMode.gamepad2.dpad_down){
-//                retractAirplane(airplaneServo);
-//            }
-            if (myOpMode.gamepad2.right_bumper){
-                openClaw(claw1);
-            }
-            if (myOpMode.gamepad2.left_bumper){
-                closeClaw(claw1);
-            }
-            if (myOpMode.gamepad2.dpad_right){
-                openClaw(claw2);
-            }
-            if (myOpMode.gamepad2.dpad_left){
-                closeClaw(claw2);
-            }
-            liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, liftMin, liftMax);
-        }
-        if (currOther == otherControls[4]) {//Michael
-//            if (myOpMode.gamepad2.square && airplaneArmed){
-//                airplanePower = airplaneMax;
-//            }else{
-//                airplanePower = 0;
-//            }
-//            if (myOpMode.gamepad2.dpad_up){
-//                shootAirplane(airplaneServo);
-//            }else if (myOpMode.gamepad2.dpad_down){
-//                retractAirplane(airplaneServo);
-//            }
-            if (myOpMode.gamepad2.dpad_up){
-                openClaw(claw1);
-            }
-            if (myOpMode.gamepad2.dpad_down){
-                closeClaw(claw1);
-            }
-            if (myOpMode.gamepad2.left_bumper){
-                openClaw(claw2);
-            }
-            if (myOpMode.gamepad2.right_bumper){
-                closeClaw(claw2);
-            }
-            liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, liftMin, liftMax);
+        if (currDriver == driverControls[5]) {//Graden
+            fieldCentric = false;
         }
     }
-
 
     public static void switchProfile(OpMode myOpMode) {
         //driver