+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/keybindMaker.py
@@ -1,10 +1,10 @@
-people = ["chase", "camden", "kian", "grady", "michael", "test"]
+people = ["chase", "camden", "kian", "grady", "michael","graden", "test"]
 keyOptions = ["dpad_down", "dpad_up", "dpad_left", "dpad_right", "left_bumper", "right_bumper", "left_stick_button",
               "right_stick_button", "circle", "triangle", "square", "cross", "left_trigger", " right_trigger",
               "left_stick_x",
               "left_stick_y", "right_stick_x", "right_stick_y"]
 driveOptions = ["fieldCentric", "robotCentric", "slowmode"]
-operatorOptions = ["openClaw", "closeClaw", "airplaneLauncher"]
+operatorOptions = ["openClaw", "closeClaw", "slidePower", "intakeOn", "flipperUp", "flipperDown", "liftPower"]
 highButtons = keyOptions[0:11]
 boundKeys = []
 person = input("Who is setting up their keybinds? ").lower()
@@ -41,49 +41,54 @@ if drive == "d":
         slowMode = input(" ").lower()
 
     print(" ")
-    print("Enter the following into the keybindings in Drivers.java for "+person[0].upper()+person[1:]+":")
+    print("Enter the following into the keybindings in Drivers.java for " + person[0].upper() + person[1:] + ":")
     print(" ")
-    print("fieldCentric = "+ driveMode)
+    print("fieldCentric = " + driveMode)
     if slowMode in highButtons:
         if "dpad" in slowMode:
-            var = slowMode[0] + slowMode[5:] +"High"
+            var = slowMode[0] + slowMode[5:] + "High"
         if "stick" in slowMode:
-            var = slowMode[0]+"Stick"+slowMode[-6:] +"High"
+            var = slowMode[0] + "Stick" + slowMode[-6:] + "High"
         if "bumper" in slowMode:
-            var = slowMode[0]+"Bumper High"
+            var = slowMode[0] + "Bumper High"
         else:
-            var = slowMode +"High"
-        print("public static boolean "+var+" = false;")
+            var = slowMode + "High"
+        print("public static boolean " + var + " = false;")
         print("// This goes above init")
-        print("if (myOpMode.gamepad1."+slowMode+" && !"+var+" && !slowModeIsOn) {")
+        print("if (myOpMode.gamepad1." + slowMode + " && !" + var + " && !slowModeIsOn) {")
         print("     slowModeIsOn = true;")
-        print("else if (myOpMode.gamepad1."+slowMode+" && !"+var+" && slowModeIsOn) {")
+        print("else if (myOpMode.gamepad1." + slowMode + " && !" + var + " && slowModeIsOn) {")
         print("     slowModeIsOn = false;")
         print("}")
-        print(var+" = myOpMode.gamepad1."+slowMode+";")
+        print(var + " = myOpMode.gamepad1." + slowMode + ";")
 if drive == "o":
     print(' ')
-    for line in operatorOptions:
-        print(line)
-    print(' ')
-    print("For the future type the first letter of the thing you want to bind except if it is a button")
     option = []
-    button = ["", "", "", "", ""]
-    for i in range(5):
+    button = ["", "", "", "", "","", "", "", ""]
+    for i in range(9):
         if i == 0:
             print("Bind the open claw 1 button")
         elif i == 1:
             print("Bind the close claw 1 button")
-        elif i == 3:
-            print("Bind the open claw 2 button")
-        elif i == 4:
-            print("Bind the close claw 2 button")
         elif i == 2:
-            print("Bind the airplane launcher button")
+            print("Bind the open claw 2 button")
+        elif i == 3:
+            print("Bind the close claw 2 button")
+        elif i == 8:
+            print("Bind the slide power button")
+        elif i == 5:
+            print("Bind the intake button")
+        elif i == 6:
+            print("Bind the flipper up button")
+        elif i == 7:
+            print("Bind the flipper down button")
+        elif i == 4:
+            print("Bind the lift power button")
+
         print("Which button do you want to bind it to? Please make it exact ")
-        print(keyOptions[0:12])
+        print(keyOptions)
         button[i] = input(" ").lower()
-        while not button[i] in keyOptions[0:12]:
+        while not button[i] in keyOptions:
             print("That is not a valid option")
             button[i] = input(" ").lower()
 
@@ -91,23 +96,25 @@ if drive == "o":
         print(" ")
 
     print(" ")
-    print("Enter the following into the keybindings in Operators.java for "+person[0].upper()+person[1:]+":")
+    print("Enter the following into the keybindings in Operators.java for " + person[0].upper() + person[1:] + ":")
     print(" ")
-    print("if (myOpMode.gamepad2."+button[2]+" && airplaneArmed){")
-    print("     airplanePower = airplaneMax;")
-    print("}else{")
-    print("     airplanePower = 0;")
+    print("if (myOpMode.gamepad2." + button[0] + "){")
+    print("     openClaw(claw1);")
     print("}")
-    print("if (myOpMode.gamepad2."+button[0]+"){")
-    print("openClaw(claw1);")
-    print("}")
-    print("if (myOpMode.gamepad2."+button[1]+"){")
+    print("if (myOpMode.gamepad2." + button[1] + "){")
     print("     closeClaw(claw1);")
     print("}")
-    print("if (myOpMode.gamepad2."+button[3]+"){")
+    print("if (myOpMode.gamepad2." + button[2] + "){")
     print("     openClaw(claw2);")
     print("}")
-    print("if (myOpMode.gamepad2."+button[4]+"){")
+    print("if (myOpMode.gamepad2." + button[3] + "){")
     print("     closeClaw(claw2);")
     print("}")
-    print("liftPower = Range.clip(myOpMode.gamepad2.left_stick_y, -liftMax, liftMax);")
+    print("liftPower = Range.clip(myOpMode.gamepad2." + button[4] + ", -liftMax, liftMax);")
+    print("slidePower = Range.clip(myOpMode.gamepad2." + button[8] + ", slideMin, slideMax);")
+    print("if (myOpMode.gamepad2." + button[6] + "){")
+    print("     ServoUtil.flipperUp(flipperServo);")
+    print("} else if (myOpMode.gamepad2." + button[7] + "){")
+    print("     ServoUtil.flipperDown(flipperServo);")
+    print('}')
+    print("intakePower = Range.clip(myOpMode.gamepad2." + button[5] + ", intakeMin, intakeMax);")