+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test.py
@@ -1,9 +1,0 @@
-read = open('keybinds.txt','r')
-lines = read.readlines()
-for line in lines:
-    if 'test' in line:
-        print('yes')
-    else:
-        print('no')
-
-read.close()